example: Fix 64bit problems in product_database_test.

+++ src/example/product_database_test.c
@@ -37,7 +37,7 @@ static void test_connect_to_product_database(void **state) {
     expect_string(connect_to_database, url, "products.abcd.org");
     expect_value(connect_to_database, port, 322);
     will_return(connect_to_database, 0xDA7ABA53);
-    assert_int_equal((int)(size_t)connect_to_product_database(), 0xDA7ABA53);
+    assert_int_equal((size_t)connect_to_product_database(), 0xDA7ABA53);
 }
 
 /* This test will fail since the expected URL is different to the URL that is
@@ -48,7 +48,7 @@ static void test_connect_to_product_database_bad_url(void **state) {
     expect_string(connect_to_database, url, "products.abcd.com");
     expect_value(connect_to_database, port, 322);
     will_return(connect_to_database, 0xDA7ABA53);
-    assert_int_equal((int)(size_t)connect_to_product_database(), 0xDA7ABA53);
+    assert_int_equal((size_t)connect_to_product_database(), 0xDA7ABA53);
 }
 
 /* This test will fail since the mock connect_to_database() will attempt to
@@ -59,7 +59,7 @@ static void test_connect_to_product_database_missing_parameter(void **state) {
 
     expect_string(connect_to_database, url, "products.abcd.org");
     will_return(connect_to_database, 0xDA7ABA53);
-    assert_int_equal((int)(size_t)connect_to_product_database(), 0xDA7ABA53);
+    assert_int_equal((size_t)connect_to_product_database(), 0xDA7ABA53);
 }
 
 int main(void) {
